modify_multi_attention/requirements.txt
torch
numpy
pyyaml
matplotlib
fightingcv